+++ src/Libraries/SmartStore.Services/Shipping/ShippingService.cs
@@ -8,11 +8,9 @@ using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Orders;
 using SmartStore.Core.Domain.Shipping;
 using SmartStore.Core.Domain.Stores;
-using SmartStore.Core.Infrastructure;
 using SmartStore.Core.Localization;
 using SmartStore.Core.Logging;
 using SmartStore.Core.Plugins;
-using SmartStore.Rules;
 using SmartStore.Services.Cart.Rules;
 using SmartStore.Services.Catalog;
 using SmartStore.Services.Common;
@@ -23,9 +21,6 @@ namespace SmartStore.Services.Shipping
 {
     public partial class ShippingService : IShippingService
     {
-		private readonly static object _lock = new object();
-		private static IList<Type> _shippingMethodFilterTypes = null;
-
 		private readonly IRepository<ShippingMethod> _shippingMethodRepository;
 		private readonly IRepository<StoreMapping> _storeMappingRepository;
 		private readonly IProductAttributeParser _productAttributeParser;
@@ -35,7 +30,6 @@ namespace SmartStore.Services.Shipping
         private readonly ShippingSettings _shippingSettings;
 		private readonly ISettingService _settingService;
 		private readonly IProviderManager _providerManager;
-		private readonly ITypeFinder _typeFinder;
 		private readonly ICommonServices _services;
         private readonly ICartRuleProvider _cartRuleProvider;
 
@@ -49,7 +43,6 @@ namespace SmartStore.Services.Shipping
             ShippingSettings shippingSettings,
 			ISettingService settingService,
 			IProviderManager providerManager,
-			ITypeFinder typeFinder,
 			ICommonServices services,
             ICartRuleProvider cartRuleProvider)
         {
@@ -62,7 +55,6 @@ namespace SmartStore.Services.Shipping
             _shippingSettings = shippingSettings;
 			_settingService = settingService;
 			_providerManager = providerManager;
-			_typeFinder = typeFinder;
 			_services = services;
             _cartRuleProvider = cartRuleProvider;
 
@@ -194,9 +186,6 @@ namespace SmartStore.Services.Shipping
 				return allMethods;
 			}
 
-			IList<IShippingMethodFilter> allFilters = null;
-			var filterRequest = new ShippingFilterRequest {	Option = request };
-
 			var activeShippingMethods = allMethods.Where(s =>
 			{
                 // Rule sets.
@@ -204,15 +193,6 @@ namespace SmartStore.Services.Shipping
                 {
                     return false;
                 }
-
-                // Shipping method filtering.
-                if (allFilters == null)
-					allFilters = GetAllShippingMethodFilters();
-
-				filterRequest.ShippingMethod = s;
-
-				if (allFilters.Any(x => x.IsExcluded(filterRequest)))
-					return false;
 
 				return true;
 			});
@@ -419,27 +399,6 @@ namespace SmartStore.Services.Shipping
 			}
             
             return result;
-        }
-
-		public virtual IList<IShippingMethodFilter> GetAllShippingMethodFilters()
-		{
-			if (_shippingMethodFilterTypes == null)
-			{
-				lock (_lock)
-				{
-					if (_shippingMethodFilterTypes == null)
-					{
-						_shippingMethodFilterTypes = _typeFinder.FindClassesOfType<IShippingMethodFilter>(ignoreInactivePlugins: true)
-							.ToList();
-					}
-				}
-            }
-
-			var shippingMethodFilters = _shippingMethodFilterTypes
-				.Select(x => EngineContext.Current.ContainerManager.ResolveUnregistered(x) as IShippingMethodFilter)
-				.ToList();
-
-			return shippingMethodFilters;
         }
 
         #endregion