+++ src/Tests/SmartStore.Services.Tests/Shipping/ShippingServiceTests.cs
@@ -8,7 +8,6 @@ using SmartStore.Core.Domain.Orders;
 using SmartStore.Core.Domain.Shipping;
 using SmartStore.Core.Domain.Stores;
 using SmartStore.Core.Events;
-using SmartStore.Core.Infrastructure;
 using SmartStore.Services.Cart.Rules;
 using SmartStore.Services.Catalog;
 using SmartStore.Services.Common;
@@ -19,7 +18,7 @@ using SmartStore.Tests;
 
 namespace SmartStore.Services.Tests.Shipping
 {
-	[TestFixture]
+    [TestFixture]
     public class ShippingServiceTests : ServiceTest
     {
         IRepository<ShippingMethod> _shippingMethodRepository;
@@ -33,7 +32,6 @@ namespace SmartStore.Services.Tests.Shipping
 		IGenericAttributeService _genericAttributeService;
         IShippingService _shippingService;
 		ISettingService _settingService;
-		ITypeFinder _typeFinder;
         ICartRuleProvider _cartRuleProvider;
 
         [SetUp]
@@ -55,7 +53,6 @@ namespace SmartStore.Services.Tests.Shipping
 
 			_genericAttributeService = MockRepository.GenerateMock<IGenericAttributeService>();
 			_settingService = MockRepository.GenerateMock<ISettingService>();
-			_typeFinder = MockRepository.GenerateMock<ITypeFinder>();
             _cartRuleProvider = MockRepository.GenerateMock<ICartRuleProvider>();
 
             _shippingService = new ShippingService(
@@ -68,7 +65,6 @@ namespace SmartStore.Services.Tests.Shipping
                 _shippingSettings,
 				_settingService, 
 				this.ProviderManager,
-				_typeFinder,
 				_services,
                 _cartRuleProvider);
         }