+++ changelog.md
@@ -7,6 +7,7 @@
 
 ### Breaking changes
 * Install\UninstallPermissions has been removed from IPermissionService. No longer needs to be called by plugins. Permissions are now automatically installed\uninstalled.
+* IShippingMethodFilter has been removed. Shipping methods can now be filtered via rule sets.
 
 ### New Features
 * #1598 Add published property to ProductTag entity.